+++ react-native/storyly-react-native/ios/STStorylyView.h
@@ -17,6 +17,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property(nonatomic, copy) RCTBubblingEventBlock onStorylyLoaded;
 @property(nonatomic, copy) RCTBubblingEventBlock onStorylyLoadFailed;
+@property(nonatomic, copy) RCTBubblingEventBlock onStorylyEvent;
 @property(nonatomic, copy) RCTBubblingEventBlock onStorylyActionClicked;
 @property(nonatomic, copy) RCTBubblingEventBlock onStorylyStoryPresented;
 @property(nonatomic, copy) RCTBubblingEventBlock onStorylyStoryDismissed;